requirements.txt
opencv-python-headless
streamlit-camera-input-live
numpy
ultralytics
Pillow